Ajout de VueModeleClientEdit
