+++ HouseMadera/HouseMadera/HouseMadera.csproj
@@ -156,7 +156,8 @@
     <Compile Include="Utilities\ConnectivityMonitor.cs" />
     <Compile Include="Utilities\RegexUtilities.cs" />
     <Compile Include="Utilites\StringToIntValidationRule.cs" />
-    <Compile Include="VueModele\VueModeleClient.cs" />
+    <Compile Include="VueModele\VueModeleClientList.cs" />
+    <Compile Include="VueModele\VueModeleClientEdit.cs" />
     <Compile Include="VueModele\VueModeleCommune.cs" />
     <Compile Include="VueModele\VueModeleChoixAdmin.cs" />
     <Compile Include="VueModele\VueModeleLogin.cs" />
